회원탈퇴 권한 설정

+++ src/main/java/com/example/prj1be/controller/MemberController.java
@@ -71,9 +71,17 @@ public class MemberController {
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login",required = false) Member login) {
         //TODO : 로그인 했는지?  안했으면 401
         //TODO : 자기 정보인지?  아니면 403
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
